1 添加失败重试 2 修复获取结果过期时间比对

+++ src/main/java/interf/ITask.java
@@ -44,7 +44,7 @@ public interface ITask {
      * 获取任务结果
      * @return
      */
-    public Object getResult(long millisecond);
+    public Object getResult(int millisecond);
 
     /**
      * 获取任务状态
